supervisor/requirements.txt
aiohttp
aiohttp_cors
python-socketio
ag-ui-protocol
google-adk
